+++ client/src/influence_ui.rs
@@ -1,4 +1,4 @@
-use crate::city_ui::CityMenu;
+use crate::ui_state::CityMenu;
 use macroquad::ui::Ui;
 use server::city_pieces::Building;
 use server::game::{Action, Game};
